requirements.txt
streamlit==1.41.0
pandas==2.2.3
openpyxl==3.1.5
Pillow==10.3.0
google-generativeai==0.8.3
requests==2.32.3
python-dotenv==1.0.1
openai==1.57.3
together==1.3.5
groq==0.13.0
fpdf==1.7.2
beautifulsoup4==4.12.2
httpx==0.28.1